Removed the Geist Fonts.
Created a new component called PageTemplate. This is the component used for each one of the gallery pages and keeps the styling and layout consistent.

Also created a search bar for the astro page to filter the images based on name or catalogue ID.

Added a function to store the theme in local storage for consistency between pages.

Used Framer Motion on the theme button instead of vanilla animation to mount and unmount the icons on the fly.

Finally, added functionality to the home page to cycle between images for the background, and calculated the height of the header in order to calculate the height of the main page without it overflowing.