Developed the header component into an almost finished state. Added the desktop menu and drop down. Also added a gallery overview option that will link to that page.
Added the equipment images and developed this page by creating EquipmentCategory.jsx, a component that is looped through to display all the equipment. Also styled this page near completion.

Applied a gradient to the background directly of each page instead of using a div.

Fixed a bug causing Fancybox to display a large overflowing image, and fixed the layout issue of the info panel.

Finally, I added the correct info for the images and equipment, and added new images in the landscape category.