Resized the woodland thumbnails.
Added a location icon and linked this to the location of where the image was taken. The icon is on all non-astro images, and I have set links for the woodland images thus far.

Added ISO, shutter speed and F stop data for woodland images.

Implemented a button and functionality to add a border to the images, both thumbnails and fancybox image.
Fancybox image styling is done via globals.css.