Overhauled the styling for a more modern and sleek look.
Removed rounded corners on favour of sharp 90degree corners.

Removed the banners at the bottom of the gallery and location categories, and put the title text over the image with a small dark gradient in the corner.

Repeated this for the hover overlay on the images except for mobile which has remained the same.

Changed the font to Manrope.