Split the imageData.json into individual files (equipment, macro, astro etc) to help with organisation. Removed the imageData.json file.
Updated the filtering to include the image date, and added date filtering buttons.

Amended the search input placeholder text to include date.

Adjusted the margin on the border button to be on both x axis instead of just the left for a more uniform look.

Finally, added 3 new images, 2 into macro and 1 into woodland.