Changed the logo - created a new logo for mobile and desktop. May need to tweak the sizing but I will test after this commit.
Also added a new local font - Satoshi and applied globally.

Also changed the page title and text to normal and light respectively (from semibold & normal).

Also removed the theme and border button from the homepage.

Swapped the border on and off display. It now shows ON when the border is on, and OFF when off for better understandability.