Changed the social logos to SVG and resized them.
Also changed the homepage to include my main logo as the center element, removed the logo in the top left on homepage, and changed the H1 to the text under the logo.

Created a new component called Logos_Icons to store all the SVG icons. Moved the social, searchbar clear, filter and other icons into here.

Tweaked the height of the theme button.

Changed the sizing of the filter help text and added further info as to how it works to make more sense.